+++ CHANGELOG.rst
@@ -7,5 +7,6 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 Unreleased
 ********
-* CIP-3440 Add repo basic necessities
+* CIP-3443 Add and test CBD device command clients, create test structure for repo, minor Makefile cleanup, add CODEOWNERS
 * CIP-3442 Add namespace deployment to repo
+* CIP-3440 Add repo basic necessities